Initial work for integrating subqueries
It reuses a lot of the previous compilation code to compile subqueries.
It breaks the compilation into a prepare and compile step. The prepare
step will read the initial outer portion of the query to determine the
global options set on it. Since the inner query needs to inherit this
information from the outside, this step happens going from outside to
in.

After options are inherited, the subquery is compiled and then the outer
query is compiled. We do not currently do anything but throw away the
result of the compilation, but that's next.